requirements.txt
gspread==5.9.0
oauth2client==4.1.3
pandas==2.0.2
peewee==3.16.2
psycopg2_binary==2.9.6
python-telegram-bot[all]==20.3